Implement serialize itmes (deserialize pending)
