+++ data/DB/genres.json
@@ -1,24 +1,14 @@
 [
   {
-    "id": 477,
-    "genre": "d",
+    "id": 828,
+    "genre": "t",
     "items": [
 
     ]
   },
   {
-    "id": 955,
-    "genre": "t",
-    "items": [
-      {
-        "item_id": 98,
-        "item_class": "MusicAlbum"
-      }
-    ]
-  },
-  {
-    "id": 554,
-    "genre": "y",
+    "id": 804,
+    "genre": "8",
     "items": [
 
     ]
