+++ data/DB/labels.json
@@ -1,12 +1,3 @@
 [
-  {
-    "id": 129,
-    "title": "y",
-    "color": "c"
-  },
-  {
-    "id": 875,
-    "title": "g",
-    "color": "y"
-  }
+
 ]
