+++ src/classes/music_album.rb
@@ -1,4 +1,5 @@
 require_relative './item'
+require_relative './genre'
 
 class MusicAlbum < Item
   attr_accessor :on_spotify
@@ -26,19 +27,10 @@ class MusicAlbum < Item
     on_spotify = gets.chomp
     on_spotify = on_spotify == 'y'
     music_album = MusicAlbum.new(on_spotify, date, genre, author, label)
-    validate_genre(genre, genres, music_album)
+    Genre.validate_genre(genre, genres, music_album)
     music_album.move_to_archive
     music_albums << music_album
+    puts genres.inspect
     puts 'Music album added'
-  end
-
-  def self.validate_genre(genre, genres, item)
-    if genres.select { |genre_item| genre_item.genre == genre }.empty?
-      new_genre = Genre.new(genre)
-      new_genre.add_item(item)
-      genres << new_genre
-    else
-      genres.select { |genre_item| genre_item.genre == genre }[0].add_item(item)
-    end
   end
 end
